further rework on repo structure
