--- focus_validator/config_validators/check_options.py
@@ -1,14 +1,0 @@
-from typing import List
-
-from pydantic import BaseModel
-
-
-class AllowNullsCheck(BaseModel):
-    allow_nulls: bool
-
-
-class ValueIn(BaseModel):
-    value_in: List[str]
-
-
-SIMPLE_CHECKS = ["check_unique"]
