--- focus_validator/config_validators/data_type_config.py
@@ -1,12 +1,0 @@
-from enum import Enum
-
-from pydantic import BaseModel
-
-
-class DataTypes(Enum):
-    STRING = "string"
-    decimal = "decimal"
-
-
-class DataTypeConfig(BaseModel):
-    data_type: DataTypes
