--- focus_validator/config_validators/dimension_value_types.py
@@ -1,7 +1,0 @@
-import pandera as pa
-
-DIMENSION_VALUE_TYPE_MAP = {
-    "InvoiceIssuer": pa.String,
-    "ResourceID": pa.String,
-    "ChargeType": pa.String,
-}
