--- focus_validator/config_validators/override_config.py
@@ -1,10 +1,0 @@
-from typing import List
-from pydantic import BaseModel
-
-
-class DimensionOverrideConfig(BaseModel):
-    skip: List[str]
-
-
-class ValidationOverrideConfig(BaseModel):
-    overrides: DimensionOverrideConfig
