+++ focus_validator/data_loaders/data_loader.py
@@ -1,6 +1,11 @@
 from focus_validator.exceptions import FocusNotImplementedError
 from .csv_data_loader import CSVDataLoader
-from .file_mime import get_file_mime_type
+import magic
+
+
+def get_file_mime_type(filename):
+    f = magic.Magic(uncompress=True)
+    return f.from_file(filename=filename)
 
 
 class DataLoader:
