--- focus_validator/data_loaders/file_mime.py
@@ -1,6 +1,0 @@
-import magic
-
-
-def get_file_mime_type(filename):
-    f = magic.Magic(uncompress=True)
-    return f.from_file(filename=filename)
