+++ focus_validator/validator.py
@@ -1,7 +1,9 @@
 from focus_validator.data_loaders import data_loader
 from focus_validator.outputter.outputter import Outputter
-from focus_validator.rules.spec_rules_loader import SpecRulesLoader
+from focus_validator.rules.spec_rules import SpecRules
 
+# TODO: Default rule_set_path here also
+# TODO: Default rules_version here also
 
 class Validator:
     def __init__(self, data_filename, rule_set_path, rules_version, output_destination, override_filename=None):
@@ -10,23 +12,20 @@ class Validator:
         self.override_filename = override_filename
 
         self.rules_version = rules_version
-        self.spec_rules_loader = SpecRulesLoader(override_filename=override_filename,
-                                                 rule_set_path=rule_set_path,
-                                                 rules_version=rules_version)
+        self.spec_rules = SpecRules(override_filename=override_filename,
+                                    rule_set_path=rule_set_path,
+                                    rules_version=rules_version)
         self.output_destination = output_destination
         self.outputter = Outputter(self.output_destination)
 
     def load(self):
         self.focus_data = data_loader.DataLoader(data_filename=self.data_filename).load()
-        self.spec_rules_loader.load()
+        self.spec_rules.load()
 
     def validate(self):
         self.load()
-        results = {}
-        for rule in self.spec_rules_loader.rules:
-            results[rule.check_id] = rule.validate(self.focus_data)
-            results[rule.check_id]['rule'] = rule
+        results = self.spec_rules.validate()
         self.outputter = self.outputter.write(results)
 
     def get_supported_versions(self):
-        return self.spec_rules_loader.supported_versions()
+        return self.spec_rules.supported_versions()
